Fix for is_1536
The function failed to recognize 1536 well plates when column numbers weren't zero padded  (i.e. `A1` vs `A01`). This is because it was checking if the number of characters in the vector of wells was equal to 4, which it wouldn't be in the case where column numbers aren't zero padded (`AA1` vs `AA01`). This change looks for wells that contain two consecutive letters.